Separate insights and standalone routes
