+++ package.json
@@ -5,22 +5,22 @@
     "author": "Red Hat, Inc.",
     "private": true,
     "dependencies": {
-        "@babel/runtime": "^7.7.2",
-        "@patternfly/patternfly": "^2.71.4",
-        "@patternfly/react-core": "^3.120.8",
+        "@babel/runtime": "^7.9.6",
+        "@patternfly/patternfly": "^2.71.6",
+        "@patternfly/react-core": "^3.153.13",
         "@redhat-cloud-services/frontend-components": "^0.0.32",
         "@redhat-cloud-services/frontend-components-utilities": "^0.0.10",
-        "@types/node": "^12.12.6",
-        "@types/react": "^16.9.11",
-        "@types/react-dom": "^16.9.4",
+        "@types/node": "^12.12.38",
+        "@types/react": "^16.9.34",
+        "@types/react-dom": "^16.9.7",
         "@types/react-router-dom": "^4.3.5",
-        "axios": "^0.19.0",
+        "axios": "^0.19.2",
         "classnames": "^2.2.5",
         "js-cookie": "^2.2.1",
-        "moment": "^2.24.0",
-        "react": "^16.11.0",
-        "react-dom": "^16.11.0",
-        "react-markdown": "^4.2.2",
+        "moment": "^2.25.3",
+        "react": "^16.13.1",
+        "react-dom": "^16.13.1",
+        "react-markdown": "^4.3.1",
         "react-redux": "^5.1.2",
         "react-router-dom": "^4.2.2",
         "react-router-hash-link": "^1.2.2",
@@ -46,47 +46,47 @@
         }
     },
     "devDependencies": {
-        "@babel/core": "^7.7.2",
+        "@babel/core": "^7.9.6",
         "@babel/plugin-proposal-class-properties": "^7.8.3",
-        "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-        "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-        "@babel/plugin-transform-runtime": "^7.6.2",
-        "@babel/preset-env": "^7.7.1",
-        "@babel/preset-flow": "^7.0.0",
-        "@babel/preset-react": "^7.7.0",
+        "@babel/plugin-proposal-object-rest-spread": "^7.9.6",
+        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+        "@babel/plugin-transform-runtime": "^7.9.6",
+        "@babel/preset-env": "^7.9.6",
+        "@babel/preset-flow": "^7.9.0",
+        "@babel/preset-react": "^7.9.4",
         "@redhat-cloud-services/frontend-components-config": "0.0.7",
         "awesome-typescript-loader": "^5.2.1",
-        "axios-mock-adapter": "^1.17.0",
+        "axios-mock-adapter": "^1.18.1",
         "babel-core": "^7.0.0-bridge.0",
-        "babel-eslint": "^10.0.3",
+        "babel-eslint": "^10.1.0",
         "babel-jest": "^24.9.0",
         "babel-plugin-dual-import": "^1.2.1",
         "babel-plugin-lodash": "^3.3.2",
         "cross-fetch": "^2.2.3",
         "css-loader": "^2.1.1",
-        "enzyme": "^3.10.0",
-        "enzyme-adapter-react-16": "^1.15.1",
-        "enzyme-to-json": "^3.4.3",
+        "enzyme": "^3.11.0",
+        "enzyme-adapter-react-16": "^1.15.2",
+        "enzyme-to-json": "^3.4.4",
         "eslint": "^5.16.0",
         "eslint-config-prettier": "^2.10.0",
         "eslint-loader": "^2.2.1",
-        "eslint-plugin-react": "^7.16.0",
-        "husky": "^4.2.3",
+        "eslint-plugin-react": "^7.19.0",
+        "husky": "^4.2.5",
         "identity-obj-proxy": "^3.0.0",
         "jest": "^24.9.0",
-        "node-sass": "^4.13.1",
+        "node-sass": "^4.14.1",
         "npm-run-all": "^4.1.5",
-        "postcss": "^7.0.21",
-        "prettier": "^1.15.3",
+        "postcss": "^7.0.29",
+        "prettier": "^1.19.1",
         "prop-types": "^15.7.2",
         "source-map-loader": "^0.2.4",
         "stylelint": "^9.10.1",
         "stylelint-config-recommended-scss": "^3.3.0",
-        "stylelint-scss": "^3.12.1",
-        "typescript": "^3.7.2",
-        "webpack": "^4.41.2",
-        "webpack-bundle-analyzer": "^3.6.0",
-        "webpack-cli": "^3.3.10"
+        "stylelint-scss": "^3.17.2",
+        "typescript": "^3.8.3",
+        "webpack": "^4.43.0",
+        "webpack-bundle-analyzer": "^3.7.0",
+        "webpack-cli": "^3.3.11"
     },
     "scripts": {
         "build": "webpack --config config/insights.prod.webpack.config.js",
