+++ src/api/index.ts
@@ -22,4 +22,3 @@ export { ActiveUserAPI } from './active-user';
 export { UserAuthType, MeType } from './response-types/user';
 export { CollectionVersionAPI } from './collection-version';
 export { MyNamespaceAPI } from './my-namespace';
-export { AuthAPI } from './auth';
