+++ src/containers/index.ts
@@ -1,0 +1,15 @@
+export { default as CertificationDashboard } from './certification-dashboard/certification-dashboard';
+export { default as CollectionContent } from './collection-detail/collection-content';
+export { default as CollectionDetail } from './collection-detail/collection-detail';
+export { default as CollectionDocs } from './collection-detail/collection-docs';
+export { default as CollectionImportLog } from './collection-detail/collection-import-log';
+export { default as EditNamespace } from './edit-namespace/edit-namespace';
+export { default as LoginPage } from './login/login';
+export { default as MyImports } from './my-imports/my-imports';
+export { default as ManageNamespace } from './namespace-detail/manage-namespace';
+export { default as PartnerDetail } from './namespace-detail/partner-detail';
+export { default as MyNamespaces } from './namespace-list/my-namespaces';
+export { default as Partners } from './namespace-list/my-namespaces';
+export { default as NotFound } from './not-found/not-found';
+export { default as Search } from './search/search';
+export { default as TokenPage } from './token/token';
