+++ src/containers/token/token.tsx
@@ -6,7 +6,7 @@ import { ClipboardCopy, Button } from '@patternfly/react-core';
 
 import { BaseHeader, Main } from '../../components';
 import { Constants } from '../../constants';
-import { AuthAPI } from '../../api';
+import { ActiveUserAPI } from '../../api';
 
 interface IState {
   tokenData: {
@@ -82,7 +82,9 @@ class TokenPage extends React.Component<RouteComponentProps, IState> {
         // the reload
         .doOffline();
     } else {
-      AuthAPI.getToken().then(result => this.setState({ tokenData: result }));
+      ActiveUserAPI.getToken().then(result =>
+        this.setState({ tokenData: result }),
+      );
     }
   }
 }
