+++ src/entry-dev.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { init } from './store';
-import App from './loaders/insights-loader';
+import App from './loaders/insights/insights-loader';
 import logger from 'redux-logger';
 import getBaseName from './utilities/getBaseName';
