+++ src/entry-standalone.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import App from './loaders/standalone-loader';
+import App from './loaders/standalone/standalone-loader';
 
 // Entrypoint for compiling the app to run in standalone mode (for all deployments
 // other than on the insights/cloud services environment)
